# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas API Contract
  version: 1.0.0
  summary: Master API contract for Atlas platform clients and services.
  description: |
    Production-grade OpenAPI 3.1 contract for Atlas.

    This specification is the single source of truth for:
    - Web frontend
    - Backend implementation
    - Mobile clients
    - Admin portal
    - External integrations

    Design guarantees:
    - RESTful, versioned API under /api/v1
    - Consistent response envelope
    - Standardized errors
    - Standardized pagination/filter/sort/search
    - Typed request and response contracts

    Note: This contract defines interfaces only. It intentionally contains no implementation logic.
  contact:
    name: Atlas API Platform Team
    email: api-platform@atlas.local
  license:
    name: Proprietary
...
servers:
  - url: https://api.atlasbank.com/api/v1
    description: Production
  - url: https://staging-api.atlasbank.com/api/v1
    description: Staging
  - url: http://localhost:3001/api/v1
    description: Local development
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Get customer profile
      description: |
        Returns personal, contact, compliance, and identity profile.

        Permissions: customer:self
        Rate limit: 600 requests / hour per user
      operationId: getProfile
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 600
        window: 1h
      responses:
        '200':
          description: Profile returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Profile]
      summary: Update customer profile
      description: |
        Updates editable customer profile fields.

        Validation:
        - Email format, phone format, postal fields, and tax metadata constraints.
        - Immutable fields rejected with validation error.

        Permissions: customer:self
        Rate limit: 120 requests / hour per user
      operationId: updateProfile
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 120
        window: 1h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePatchRequest'
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /profile/preferences:
    get:
      tags: [Profile]
      summary: Get profile preferences
      description: |
        Returns theme, language, timezone, accessibility, and notification preferences.

        Permissions: customer:self
        Rate limit: 600 requests / hour per user
      operationId: getProfilePreferences
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 600
        window: 1h
      responses:
        '200':
          description: Preferences returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfilePreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Profile]
      summary: Update profile preferences
      description: |
        Updates customer preferences.

        Validation:
        - Enumerated values for language/theme.
        - Currency and timezone must be supported values.

        Permissions: customer:self
        Rate limit: 120 requests / hour per user
      operationId: patchProfilePreferences
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 120
        window: 1h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePreferencesPatchRequest'
      responses:
        '200':
          description: Preferences updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfilePreferences'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /profile/security:
    get:
      tags: [Profile]
      summary: Get profile security settings
      description: |
        Returns security controls, trusted devices, and active sessions.

        Permissions: customer:self
        Rate limit: 300 requests / hour per user
      operationId: getProfileSecurity
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 300
        window: 1h
      responses:
        '200':
          description: Security profile returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileSecurity'
    patch:
      tags: [Profile]
      summary: Update profile security settings
      description: |
        Updates profile security controls such as MFA and session settings.

        Permissions: customer:self
        Rate limit: 60 requests / hour per user
      operationId: patchProfileSecurity
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 60
        window: 1h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSecurityPatchRequest'
      responses:
        '200':
          description: Security settings updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileSecurity'
        '400':
          $ref: '#/components/responses/ValidationError'
  /profile/activity:
    get:
      tags: [Profile]
      summary: List profile activity
      description: |
        Returns profile-related activity feed.

        Supports standardized filtering, sorting, search, and pagination.

        Permissions: customer:self
        Rate limit: 600 requests / hour per user
      operationId: listProfileActivity
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 600
        window: 1h
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Activity items returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ActivityItem'
                      meta:
                        $ref: '#/components/schemas/CursorMeta'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for authenticated customer/admin APIs.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token cookie used for token refresh endpoint.
    sessionCookie:
      type: apiKey
      in: cookie
      name: atlas_session
      description: Session cookie for session-based flows.
    oauth2:
      type: oauth2
      description: Future OAuth 2.1 support for external integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.atlasbank.com/oauth2/authorize
          tokenUrl: https://auth.atlasbank.com/oauth2/token
          scopes:
            atlas.read: Read access to Atlas resources.
            atlas.write: Write access to Atlas resources.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Cursor token for forward pagination.
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      description: Maximum items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Sort:
      name: sort
      in: query
      description: Field name used for sorting.
      schema:
        type: string
    Order:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    Search:
      name: search
      in: query
      description: Full-text search term.
      schema:
        type: string
        minLength: 1
        maxLength: 120
    FromDate:
      name: fromDate
      in: query
      description: Inclusive start datetime (ISO-8601).
      schema:
        type: string
        format: date-time
    ToDate:
      name: toDate
      in: query
      description: Inclusive end datetime (ISO-8601).
      schema:
        type: string
        format: date-time
...
  responses:
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validation:
              value:
                success: false
                data: null
                meta:
                  traceId: trc_4001
                  timestamp: '2026-07-09T13:22:00Z'
                error:
                  type: validation_error
                  code: VAL_001
                  message: One or more fields failed validation.
                  details:
                    - field: email
                      reason: Invalid email format.
    Unauthorized:
      description: Authentication failed or missing credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but lacking required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, meta, error]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ErrorObject'
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            success:
              const: false
            data:
              type: 'null'
            error:
              $ref: '#/components/schemas/ErrorObject'
    ResponseMeta:
      type: object
      required: [traceId, timestamp]
      properties:
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        rateLimit:
          $ref: '#/components/schemas/RateLimitMeta'
    PaginationMeta:
      type: object
      properties:
        mode:
          type: string
          enum: [cursor, offset]
        cursor:
          type: object
          properties:
            next:
              type: string
            previous:
              type: string
            hasMore:
              type: boolean
        offset:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
    RateLimitMeta:
      type: object
      properties:
        limit:
          type: integer
        remaining:
          type: integer
        resetAt:
          type: string
          format: date-time
    ErrorObject:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          enum:
            - validation_error
            - authentication_error
            - authorization_error
            - business_rule_error
            - rate_limit_error
            - server_error
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
        traceId:
          type: string
...
    CursorMeta:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                mode:
                  const: cursor
                cursor:
                  type: object
                  required: [hasMore]
                  properties:
                    next:
                      type: string
                    previous:
                      type: string
                    hasMore:
                      type: boolean
...
    Session:
      type: object
      required: [id, device, ipAddress, createdAt, expiresAt, isCurrent]
      properties:
        id:
          type: string
          format: uuid
        device:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
    Profile:
      type: object
      required: [id, firstName, lastName, email, verificationStatus]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
        employment:
          $ref: '#/components/schemas/Employment'
        taxInfo:
          $ref: '#/components/schemas/TaxInfo'
        verificationStatus:
          type: string
          enum: [verified, pending, under_review]
    ProfilePatchRequest:
      type: object
      additionalProperties: false
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          minLength: 7
          maxLength: 20
        address:
          $ref: '#/components/schemas/Address'
        employment:
          $ref: '#/components/schemas/Employment'
        taxInfo:
          $ref: '#/components/schemas/TaxInfo'
    ProfilePreferences:
      type: object
      required: [language, currency, timezone, theme]
      properties:
        language:
          type: string
          enum: [en, es]
        currency:
          type: string
          minLength: 3
          maxLength: 3
        timezone:
          type: string
        theme:
          type: string
          enum: [light, dark, system]
        accessibility:
          type: object
          properties:
            reducedMotion:
              type: boolean
            highContrast:
              type: boolean
            largeText:
              type: boolean
            screenReaderOptimized:
              type: boolean
        notificationPreferences:
          $ref: '#/components/schemas/NotificationPreferences'
    ProfilePreferencesPatchRequest:
      type: object
      additionalProperties: false
      properties:
        language:
          type: string
          enum: [en, es]
        currency:
          type: string
          minLength: 3
          maxLength: 3
        timezone:
          type: string
        theme:
          type: string
          enum: [light, dark, system]
        accessibility:
          type: object
        notificationPreferences:
          type: object
    ProfileSecurity:
      type: object
      required: [mfaEnabled, loginAlertsEnabled, connectedDevices, activeSessions]
      properties:
        mfaEnabled:
          type: boolean
        mfaMethod:
          type: string
        loginAlertsEnabled:
          type: boolean
        biometricUnlockEnabled:
          type: boolean
        sessionTimeoutMinutes:
          type: integer
        connectedDevices:
          type: array
          items:
            $ref: '#/components/schemas/ConnectedDevice'
        activeSessions:
          type: array
          items:
            $ref: '#/components/schemas/Session'
    ProfileSecurityPatchRequest:
      type: object
      additionalProperties: false
      properties:
        mfaEnabled:
          type: boolean
        mfaMethod:
          type: string
        loginAlertsEnabled:
          type: boolean
        biometricUnlockEnabled:
          type: boolean
        sessionTimeoutMinutes:
          type: integer
          minimum: 5
          maximum: 240
    ActivityItem:
      type: object
      required: [id, type, title, description, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
    Address:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
          minLength: 2
          maxLength: 2
    Employment:
      type: object
      properties:
        employerName:
          type: string
        jobTitle:
          type: string
        status:
          type: string
        annualIncomeRange:
          type: string
    TaxInfo:
      type: object
      properties:
        taxResidencyCountry:
          type: string
          minLength: 2
          maxLength: 2
        taxIdMasked:
          type: string
        w9OnFile:
          type: boolean
...
    NotificationPreferences:
      type: object
      properties:
        emailEnabled:
          type: boolean
        smsEnabled:
          type: boolean
        pushEnabled:
          type: boolean
        transactionAlerts:
          type: boolean
        securityAlerts:
          type: boolean
        marketingEmails:
          type: boolean
        loginAlerts:
          type: boolean
        balanceAlerts:
          type: boolean
        transferAlerts:
          type: boolean
...